+++ dotfiles/hypr/layout.lua
@@ -1,5 +1,0 @@
-hl.config({
-    dwindle = {
-        preserve_split = true, -- You probably want this
-    },
-})